requirements.txt
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.9.3
google-cloud-functions==1.16.3
google-cloud-run==0.10.5
google-cloud-scheduler==2.13.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-re2==1.1
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
functions-framework==3.*
google-cloud-bigquery
requests
openmeteo-requests
requests-cache 
retry-requests
numpy
pandas
fsspec
gcsfs
